% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.r
\name{im_add_evidence}
\alias{im_add_evidence}
\title{Add evidence to a diagnosis request object}
\usage{
im_add_evidence(diag_obj, id, choice = c("present", "absent", "unknown"))
}
\arguments{
\item{diag_obj}{diagnosis request object started with \code{im_start_dignosis()}.}

\item{id}{a symptom}
}
\description{
Meant to be used in a "pipe" and will build the JSON array for POSTing to the
Infermedica API automagically in the background. Chain multiple calls together
to add additional evidence components.
}
\details{
TODO: Add support for \code{observed_at}, \code{extras} and \code{evaluated_at}
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_API_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}
\examples{
im_start_diagnosis() \%>\%
  im_add_patient_info(29, "male") \%>\%
  im_add_evidence("s_21", "present") \%>\%
  im_get_diagnosis()
}

