% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.r
\name{im_add_patient_info}
\alias{im_add_patient_info}
\title{Add patient info to a diagnosis request object}
\usage{
im_add_patient_info(diag_obj, age, sex)
}
\arguments{
\item{diag_obj}{diagnosis request object started with \code{im_start_dignosis()}.}

\item{age}{age of subject}

\item{sex}{sex of subject}
}
\description{
Meant to be used in a "pipe"
}
\details{
TODO: Add support for \code{observed_at}, \code{extras} and \code{evaluated_at}
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_API_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}
\examples{
im_start_diagnosis() \%>\%
  im_add_patient_info(29, "male") \%>\%
  im_add_evidence("s_21", "present") \%>\%
  im_get_diagnosis()
}

