% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.r
\name{im_get_condition}
\alias{im_get_condition}
\title{Retrieve info for a specific symptom}
\usage{
im_get_condition(condition_id)
}
\arguments{
\item{condition_id}{valid Infermedica condition id (use \code{im_list_conditions()} to
get a list of valid ids)}
}
\description{
Retrieve info for a specific symptom
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

