% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_tests.r, R/risk_factors.r
\name{im_get_risk_factors}
\alias{im_get_risk_factors}
\title{Retrieve info for a specific lab test}
\usage{
im_get_risk_factors(risk_factor_id)

im_get_risk_factors(risk_factor_id)
}
\arguments{
\item{risk_factor_id}{valid Infermedica risk factor id (use \code{im_list_risk_factors()} to
get a list of valid ids)}

\item{lab_test_id}{valid Infermedica risk factor id (use \code{im_list_lab_tests()} to
get a list of valid ids)}
}
\description{
Retrieve info for a specific lab test

Retrieve info for a specific risk factor
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.

The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

