% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptoms.r
\name{im_get_symptom}
\alias{im_get_symptom}
\title{Retrieve info for a specific symptom}
\usage{
im_get_symptom(symptom_id)
}
\arguments{
\item{symptom_id}{valid Infermedica symptom id (use \code{im_list_symptoms()} to
get a list of valid ids)}
}
\description{
Retrieve info for a specific symptom
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

