% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infermedica.r
\name{im_info}
\alias{im_info}
\title{Retrieve Infermedica API info}
\usage{
im_info()
}
\description{
Retrieve Infermedica API info
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

