% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.r
\name{im_list_conditions}
\alias{im_list_conditions}
\title{Retrieve list of conditions from Infermedica API}
\usage{
im_list_conditions()
}
\description{
Retrieve list of conditions from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

