% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_tests.r
\name{im_list_lab_tests}
\alias{im_list_lab_tests}
\title{Retrieve list of lab tests from Infermedica API}
\usage{
im_list_lab_tests()
}
\description{
Retrieve list of lab tests from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

