% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factors.r
\name{im_list_risk_factors}
\alias{im_list_risk_factors}
\title{Retrieve list of risk factors from Infermedica API}
\usage{
im_list_risk_factors()
}
\description{
Retrieve list of risk factors from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

