% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptoms.r
\name{im_list_symptoms}
\alias{im_list_symptoms}
\title{Retrieve list of symptoms from Infermedica API}
\usage{
im_list_symptoms()
}
\description{
Retrieve list of symptoms from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}

