% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{im_lookup}
\alias{im_lookup}
\title{Returns a single observation matching given phrase from Infermedica API}
\usage{
im_lookup(phrase, sex = NULL)
}
\arguments{
\item{phrase}{search terms/phrase}

\item{sex}{male/female or NULL}
}
\description{
Returns a single observation matching given phrase from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}
\examples{
im_lookup("search phrase")
}

