% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{im_search}
\alias{im_search}
\title{Returns list of observations matching the given phrase from Infermedica API}
\usage{
im_search(phrase, sex = NULL, type = NULL, max_results = 8)
}
\arguments{
\item{phrase}{search terms/phrase}

\item{sex}{male/female or NULL}

\item{type}{character vector or NULL. If not NULL or a single string, should be
a character vector composed of "symptom", "risk_factor", "lab_test"}

\item{max_results}{limit the number of search results (default == 8)}
}
\description{
Returns list of observations matching the given phrase from Infermedica API
}
\note{
The package expects \code{INFERMEDICA_APP_ID} and \code{INFERMEDICA_APP_KEY} to
    be in the environment. The easiest way to do this is to add entries to
    \code{~/.Renviron} and restart R.
}
\examples{
im_search("bleeding", "male", type="symptom")
}

