% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infermedica-package.R
\name{infermedica-exports}
\alias{\%>\%}
\alias{infermedica-exports}
\title{infermedica exported operators}
\description{
The following functions are imported and then re-exported
from the infermedica package to enable use of the magrittr
pipe operator with no additional library calls
}

