% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infermedica-package.R
\docType{package}
\name{infermedica}
\alias{infermedica}
\alias{infermedica-package}
\title{R API Interface to the Infermedica API}
\description{
R API Interface to the Infermedica API
}
\author{
Bob Rudis (bob@rud.is)
}

